+++ python/docs/api/uagents/config.md
@@ -4,7 +4,7 @@
 
 
 
-#### parse_endpoint_config[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/config.py#L61)
+#### parse_endpoint_config[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/config.py#L52)
 ```python
 def parse_endpoint_config(
         endpoint: str | list[str] | dict[str, dict] | None,
@@ -27,11 +27,11 @@ Parse the user-provided endpoint configuration.
 
 **Returns**:
 
-- `[List[AgentEndpoint]` - The parsed endpoint configuration.
+- `[list[AgentEndpoint]` - The parsed endpoint configuration.
 
 
 
-#### parse_agentverse_config[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/config.py#L116)
+#### parse_agentverse_config[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/config.py#L106)
 ```python
 def parse_agentverse_config(
         config: str | dict[str, str] | None = None) -> AgentverseConfig
