+++ python/docs/api/uagents/dispatch.md
@@ -4,7 +4,7 @@
 
 
 
-## Sink Objects[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/dispatch.py#L13)
+## Sink Objects[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/dispatch.py#L14)
 
 ```python
 class Sink(ABC)
@@ -14,7 +14,7 @@ Abstract base class for sinks that handle messages.
 
 
 
-## Dispatcher Objects[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/dispatch.py#L31)
+## Dispatcher Objects[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/dispatch.py#L32)
 
 ```python
 class Dispatcher()
