--- python/docs/api/uagents/models.md
@@ -1,14 +1,0 @@
-
-
-# src.uagents.models
-
-
-
-## ErrorMessage Objects[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/models.py#L36)
-
-```python
-class ErrorMessage(Model)
-```
-
-Error message model
-
