+++ python/docs/api/uagents/registration.md
@@ -4,7 +4,7 @@
 
 
 
-#### coerce_metadata_to_str[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/registration.py#L83)
+#### coerce_metadata_to_str[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/registration.py#L49)
 ```python
 def coerce_metadata_to_str(
         metadata: dict[str, Any] | None
@@ -15,7 +15,7 @@ Step through the metadata and convert any non-string values to strings.
 
 
 
-#### extract_geo_metadata[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/registration.py#L100)
+#### extract_geo_metadata[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/registration.py#L66)
 ```python
 def extract_geo_metadata(
         metadata: dict[str, Any] | None) -> dict[str, Any] | None
@@ -25,7 +25,7 @@ Extract geo-location metadata from the metadata dictionary.
 
 
 
-#### almanac_api_post[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/registration.py#L109)
+#### almanac_api_post[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/registration.py#L75)
 ```python
 async def almanac_api_post(url: str,
                            data: BaseModel,
@@ -38,7 +38,7 @@ Send a POST request to the Almanac API.
 
 
 
-## LedgerBasedRegistrationPolicy Objects[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/registration.py#L268)
+## LedgerBasedRegistrationPolicy Objects[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/registration.py#L234)
 
 ```python
 class LedgerBasedRegistrationPolicy(AgentRegistrationPolicy)
@@ -46,7 +46,7 @@ class LedgerBasedRegistrationPolicy(AgentRegistrationPolicy)
 
 
 
-#### check_contract_version[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/registration.py#L326)
+#### check_contract_version[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/registration.py#L292)
 ```python
 def check_contract_version() -> None
 ```
@@ -56,7 +56,7 @@ if it is different from the supported version.
 
 
 
-#### register[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/registration.py#L340)
+#### register[↗](https://github.com/fetchai/uAgents/blob/main/python/src/uagents/registration.py#L306)
 ```python
 async def register(agent_identifier: str,
                    identity: Identity,
